requirements.txt
ultralytics
opencv-python
tqdm
rich
PyQt5==5.15.10
pyqtgraph==0.13.7
PyYAML==6.0.1
PyOpenGL==3.1.7
PyOpenGL_accelerate==3.1.7
